+++ formio/models/formio_form.py
@@ -117,7 +117,7 @@ class Form(models.Model):
     show_state = fields.Boolean("Show State")
     show_id = fields.Boolean("Show ID")
     show_uuid = fields.Boolean("Show UUID")
-    show_user_metadata = fields.Boolean("Show User Metadata")
+    show_user_metadata = fields.Boolean(string="Show User Metadata")
     iframe_resizer_body_margin = fields.Char(
         "iFrame Resizer bodyMargin",
         related="builder_id.iframe_resizer_body_margin",